requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
asyncio-throttle==1.0.2
pydantic==2.5.0
httpx==0.25.2
schedule==1.2.0
python-jose[cryptography]==3.3.0